[BUGFIX] memalign was fucked on 64-bit

+++ newlib/libc/stdlib/_mallocr.c
@@ -3095,7 +3095,7 @@ Void_t* mEMALIGn(RARG alignment, bytes) RDECL size_t alignment; size_t bytes;
       this is always possible.
     */
 
-    brk = (char*)mem2chunk(((unsigned long)(m + alignment - 1)) & -alignment);
+    brk = (char*)mem2chunk(((uintptr_t)(m + alignment - 1)) & -alignment);
     if ((long)(brk - (char*)(p)) < (long)MINSIZE) brk = brk + alignment;
 
     newp = (mchunkptr)brk;
